Create separate tables class
